requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
certifi==2026.2.25
charset-normalizer==3.4.7
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.3.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
fastapi==0.125.0
filelock==3.28.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.14.0
hf-xet==1.4.3
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.1
huggingface_hub==1.11.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
jsonpatch==1.33
jsonpointer==3.1.1
kubernetes==35.0.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==0.28.1
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
overrides==7.7.0
packaging==23.2
posthog==7.12.0
propcache==0.4.1
protobuf==6.33.6
pulsar-client==3.11.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.20.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.50.0
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.5.2
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.1
